+++ src/main/java/com/example/hrproject/Repositories/CompetenceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CompetenceRepository extends JpaRepository<Competence, Long> {
-    List<Competence> findAllByCollaborateur_Matricule(int matricule);
+    List<Competence> findAllByCollaborateur_Matricule(Long matricule);
 
 }
