+++ src/main/java/com/example/hrproject/Repositories/DiplomeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface DiplomeRepository extends JpaRepository<Diplome, Long> {
-    List<Diplome> findAllByCollaborateur_Matricule(int matricule);
+    List<Diplome> findAllByCollaborateur_Matricule(Long matricule);
 }
