+++ src/main/java/com/example/hrproject/Services/CompetenceService.java
@@ -1,12 +1,7 @@
 package com.example.hrproject.Services;
 
-import com.example.hrproject.Entities.Collaborateur;
 import com.example.hrproject.Entities.Competence;
-import com.example.hrproject.Entities.Diplome;
-import com.example.hrproject.Repositories.CollaborateurRepository;
 import com.example.hrproject.Repositories.CompetenceRepository;
-import com.example.hrproject.Repositories.DiplomeRepository;
-import com.example.hrproject.modals.CollabDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,7 +12,7 @@ public class CompetenceService {
 
     private final CompetenceRepository competenceRepository;
 
-    public List<Competence> getComptences(int matricule) {
+    public List<Competence> getComptences(Long matricule) {
         return competenceRepository.findAllByCollaborateur_Matricule(matricule);
     }
     @Autowired
